+++ draft-packages/hierarchical-menu/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.45](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@2.0.44...@kaizen/draft-hierarchical-menu@2.0.45) (2021-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
 ## [2.0.44](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@2.0.43...@kaizen/draft-hierarchical-menu@2.0.44) (2021-03-25)
 
 **Note:** Version bump only for package @kaizen/draft-hierarchical-menu