+++ draft-packages/search-box/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.28](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.27...@kaizen/draft-search-box@1.1.28) (2021-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
 ## [1.1.27](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.26...@kaizen/draft-search-box@1.1.27) (2021-03-25)
 
 **Note:** Version bump only for package @kaizen/draft-search-box