+++ draft-packages/select/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.13.5...@kaizen/draft-select@1.13.6) (2021-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
 ## [1.13.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.13.4...@kaizen/draft-select@1.13.5) (2021-03-25)
 
 **Note:** Version bump only for package @kaizen/draft-select