+++ draft-packages/tooltip/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.6.3...@kaizen/draft-tooltip@2.6.4) (2021-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
 ## [2.6.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.6.2...@kaizen/draft-tooltip@2.6.3) (2021-03-25)
 
 **Note:** Version bump only for package @kaizen/draft-tooltip